replace refs to 'master' branch with main (#1795)
the CONTRIBUTING docs, while they have a note on language, still
refer to a 'master' branch which doesn't exist in the repo. This
simply replaces 'master' with 'main' in CONTRIBUTING.md and
copies that file to .github/CONTRIBUTING.md too. There is one
more ref to 'master' in the main readme.md but it's for coveralls
and replacing 'main' in it doesn't fix the badge so I'm not sure
whether to change it