Groups with Roles feature (#1859)
Groups with Roles feature

Introduce ability to add roles to groups. Group members inherit the permissions attached to the group's roles.

Notable changes:
- Several new Roles for Work, Collection, and User management
- Role permission integration with Hyrax permission logic (Sipity, access controls, CanCan, etc.)
- Hyku::Group replaced with already existing Hyrax::Group
- Easily assign roles to individual users
- Extended UI
- Both pre-configured and customizable groups
- Scripts to migrate existing Hyku applications
- More

Documentation for this feature can be found in the Hyku GitHub Wiki here:

https://github.com/samvera/hyku/wiki/Groups-with-Roles-Feature

This work is a joint contribution of PALNI and PALCI's Hyku for Consortia project.