🐛 Adding conditional when Devise is non-registerable
Prior to this commit, if we remove the User's `devise registerable`
configuration we would encounter an error with the path methods not
existing.  With this change, we allow for disabling of the
registerable feature; in a Devise friendly manner.

We could look to the below view and extract a helper method for this.
But this seems to be a path of least problem.

From `app/views/devise/shared/_links.html.erb`:

```
<%- if devise_mapping.registerable? && controller_name != 'registrations' && !Account.global_tenant? && current_account.try(:allow_signup) %>
  <%= link_to t(".sign_up"), new_registration_path(resource_name) %><br />
<% end -%>
```

The `devise_mapping` method is `Devise.mappings[<key>]`; and in this
case that key is `:user`.  See
https://github.com/heartcombo/devise/blob/ec0674523e7909579a5a008f16fb9fe0c3a71712/app/controllers/devise_controller.rb#L60

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/492