⚙️ Add bundle as part of docker-compose
In working on a new HykuKnapsack, I encountered a problem where the
`db-migrate-seed.sh` attempted to call `bundle exec rails db:create` and
the command failed because `rails` was not installed.

With this change, we'll ensure that we run `bundle` each time we
initialize the application.  With newer versions of bundler that appears
to be a relatively fast command (a few seconds).  In doing so, we ensure
that the first time we bring up the docker environment, we are building
those dependencies. (Note: When we build those dependencies, it will
take longer than a few seconds, but going forward those should be cached)

A word of caution, if you observe that this addition adds several
minutes to your Hyku build, we need to revisit this approach.  In prior
versions of bundler, we've observed that adding bundle as part of the
build would typically add 2 to 4 minutes each time you run `docker
compose up`.

Related to:

- https://github.com/samvera/hyku/pull/2004