🎁 Superadmins can activate users
This commit will allow superadmins to activate users from the dashboard.
The default password for an activated user is set to `password`.  The
activate button shows up for users who have permissions to destroy
users.  The assumption is that superadmins will be the only ones with
this permission.

Also the locales have been updated to include the new strings.