remove duplicate JAVA_OPTS env var (#2038)
The same variable is defined two lines above and has a value, so I assume that is the correct one.

The commit that added this duplicate (`ff2888fa`) mentions various improvements (Bulkrax, Docker build, etc). Does removing the empty var break any of that? Should the var with a value be removed instead?